Books online search refactoring
