+++ app/src/main/java/dev/zezula/books/data/BookSearchResultsRepositoryImpl.kt
@@ -7,6 +7,7 @@ import dev.zezula.books.data.model.book.SearchBookResultEntity
 import dev.zezula.books.data.model.book.toBookEntity
 import dev.zezula.books.data.model.book.asExternalModel
 import dev.zezula.books.data.source.db.BookDao
+import dev.zezula.books.data.source.db.BookSearchResultDao
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import timber.log.Timber
@@ -15,10 +16,11 @@ import java.util.UUID
 
 class BookSearchResultsRepositoryImpl(
     private val bookDao: BookDao,
+    private val bookSearchResultDao: BookSearchResultDao,
 ) : BookSearchResultsRepository {
 
-    override fun getAllSearchResultsStream(): Flow<List<Book>> {
-        return bookDao.getAllSearchResultBooksStream().map {
+    override fun getAllSearchResultsFlow(): Flow<List<Book>> {
+        return bookSearchResultDao.getAllSearchResultBooksFlow().map {
             it.map(BookEntity::asExternalModel)
         }
     }
@@ -30,11 +32,11 @@ class BookSearchResultsRepositoryImpl(
             dateAdded = LocalDateTime.now().toString(),
         )
         bookDao.insertBook(bookEntity)
-        bookDao.addToSearchBookResults(SearchBookResultEntity(bookId = bookId))
+        bookSearchResultDao.insertSearchBookResults(SearchBookResultEntity(bookId = bookId))
     }
 
     override suspend fun deleteAllSearchResults() {
         Timber.d("deleteAllSearchBookResults()")
-        bookDao.deleteAllSearchResults()
+        bookSearchResultDao.deleteAllSearchResults()
     }
 }
