+++ app/src/main/java/dev/zezula/books/di/AppModule.kt
@@ -151,6 +151,10 @@ val appModule = module {
     }
     single {
         val database = get<AppDatabase>()
+        database.bookSearchResultDao()
+    }
+    single {
+        val database = get<AppDatabase>()
         database.noteDao()
     }
     single {
@@ -198,7 +202,7 @@ val appModule = module {
     // Repositories
     single<BooksRepository> { BooksRepositoryImpl(get(), get(), get()) }
     single<BookSuggestionsRepository> { BookSuggestionsRepositoryImpl(get(), get()) }
-    single<BookSearchResultsRepository> { BookSearchResultsRepositoryImpl(get()) }
+    single<BookSearchResultsRepository> { BookSearchResultsRepositoryImpl(get(), get()) }
     single<UserLibraryRepository> { UserLibraryRepositoryImpl(get(), get()) }
     single<NotesRepository> { NotesRepositoryImpl(get()) }
     single<UserRepository> { UserRepositoryImpl() }
